Update hash to handle when hash_map expands
Now the hash function should actually generate hash_codes that go out
into the new buckets when the hash_map expands.

It doesn't make sense to me that we would check if the hash_code is
greater than the current size if it's % capacity. That should be
impossible, so I removed that from the error handling.